refactor: HttpResponse 클래스명을 ResponseGenerator로 수정
- 클래스명을 역할에 맞게 ResponseGenerator로 수정한다
- 리다이렉트를 처리하기 위한 response302Header()를 추가한다
